requirements.txt
GitPython>=3.1.0
tabulate>=0.6.0
